// src/components/molecules/mainMenuBar/nav/types.ts
export interface TopNavProps {
  ariaDetails?: any
  $fontFamily?: string
  $fontColor?: string
  $hoverColor?: string
  linkFontFamily?: string
  linkFontSize?: string
  linkFontWeight?: string
  linkLetterSpacing?: string
  linkPadding?: string
  pages?: any
  imgStyles?: any
  navBarImgSource?: string
  navBarImgAlt?: string
  navBarImgStyles?: any
}

export interface StyledNavBarProps {
  $aPadding?: string
  $fontColor?: string
  $fontFamily?: string
  $fontSize?: string
  $imgStyles?: any
  $fontWeight?: string
  $letterSpacing?: string
  $textTransform?: string
  $hoverColor?: string
  ariaDetails?: any
  navBarBorderBottom?: string
  navBarBoxShadow?: string
  navBarImg?: any
}

export interface NavBarLinkProps {
  ariaDetails?: string
  fontFamily?: string
  pages: any
  linkcolor?: string
  linkhovercolor?: string
  imgStyle?: any
}